+++ admin/menus.php
@@ -16,19 +16,19 @@ $rs_menu = new Menu;
 	switch($action) {
 		case 'create':
 			// Create a new menu
-			userHasPrivilege($_SESSION['role'], 'can_create_menus') ? $rs_menu->createMenu() : redirect('menus.php');
+			userHasPrivilege($session['role'], 'can_create_menus') ? $rs_menu->createMenu() : redirect('menus.php');
 			break;
 		case 'edit':
 			// Edit an existing menu
-			userHasPrivilege($_SESSION['role'], 'can_edit_menus') ? $rs_menu->editMenu($id) : redirect('menus.php');
+			userHasPrivilege($session['role'], 'can_edit_menus') ? $rs_menu->editMenu($id) : redirect('menus.php');
 			break;
 		case 'delete':
 			// Delete an existing menu
-			userHasPrivilege($_SESSION['role'], 'can_delete_menus') ? $rs_menu->deleteMenu($id) : redirect('menus.php');
+			userHasPrivilege($session['role'], 'can_delete_menus') ? $rs_menu->deleteMenu($id) : redirect('menus.php');
 			break;
 		default:
 			// List all menus
-			userHasPrivilege($_SESSION['role'], 'can_view_menus') ? $rs_menu->listMenus() : redirect('index.php');
+			userHasPrivilege($session['role'], 'can_view_menus') ? $rs_menu->listMenus() : redirect('index.php');
 	}
 	?>
 </div>