+++ admin/widgets.php
@@ -16,19 +16,19 @@ $rs_widget = new Widget;
 	switch($action) {
 		case 'create':
 			// Create a new widget
-			userHasPrivilege($_SESSION['role'], 'can_create_widgets') ? $rs_widget->createWidget() : redirect('widgets.php');
+			userHasPrivilege($session['role'], 'can_create_widgets') ? $rs_widget->createWidget() : redirect('widgets.php');
 			break;
 		case 'edit':
 			// Edit an existing widget
-			userHasPrivilege($_SESSION['role'], 'can_edit_widgets') ? $rs_widget->editWidget($id) : redirect('widgets.php');
+			userHasPrivilege($session['role'], 'can_edit_widgets') ? $rs_widget->editWidget($id) : redirect('widgets.php');
 			break;
 		case 'delete':
 			// Delete an existing widget
-			userHasPrivilege($_SESSION['role'], 'can_delete_widgets') ? $rs_widget->deleteWidget($id) : redirect('widgets.php');
+			userHasPrivilege($session['role'], 'can_delete_widgets') ? $rs_widget->deleteWidget($id) : redirect('widgets.php');
 			break;
 		default:
 			// List all widgets
-			userHasPrivilege($_SESSION['role'], 'can_view_widgets') ? $rs_widget->listWidgets() : redirect('index.php');
+			userHasPrivilege($session['role'], 'can_view_widgets') ? $rs_widget->listWidgets() : redirect('index.php');
 	}
 	?>
 </div>